Wrong OBO id workaround
Instead of trying to parse the ontology prefix and term id from the OBO id which is currently broken https://github.com/EBISPOT/OLS/issues/401, we query the OLS API to use the base uris to obtain the id, and try to get the prefix from the set of preferredPrefixes. Its fine now, but in the rewrite we should stablise the code for curie generation more.